+++ opti/problems/data/cake.csv
@@ -1,0 +1,51 @@
+wheat_flour,spelt_flour,sugar,chocolate,nuts,carrot,calories,taste,browning
+0.21,0,0.26,0.27,0.14,0.12,436.7738521,3.497981406,0.758500085
+0.29,0,0.29,0.26,0,0.16,395.0992445,0.776140092,0.749857272
+0,0.37,0.18,0.26,0.19,0,465.1029138,4.055489634,1.352
+0.2,0,0.15,0.32,0.14,0.19,339.3823768,3.197262889,0.826815588
+0.13,0.2,0.26,0.32,0,0.09,398.3182445,0.904554414,1.138354875
+0,0.37,0.34,0.22,0.07,0,459.9732096,2.09344882,1.312
+0.29,0,0.32,0.28,0,0.11,354.3485434,0.912223642,0.802974807
+0.34,0,0.15,0.34,0.17,0,484.6434736,3.864211767,0.815400523
+0.13,0.14,0.34,0.28,0.11,0,450.8474392,3.00055058,1.00846061
+0.1,0.23,0.19,0.2,0.09,0.19,381.2843246,1.951460934,1.069707914
+0.11,0.22,0.31,0.22,0,0.14,346.3180133,0.800373036,1.074004817
+0.25,0.12,0.3,0.2,0.13,0,459.2451023,3.181324487,0.888057674
+0.16,0.19,0.19,0.3,0.16,0,477.7458525,3.658017321,1.112738931
+0.12,0.16,0.34,0.18,0.2,0,462.862761,4.458175209,0.942103402
+0,0.33,0.34,0.24,0,0.09,377.3769182,0.902841949,1.268
+0.37,0,0.35,0.17,0.11,0,418.6744289,2.88084879,0.683555134
+0.25,0,0.16,0.21,0.19,0.19,366.4131553,3.682613224,0.700691489
+0.15,0.24,0.18,0.23,0.2,0,431.3347645,4.188142987,1.121733587
+0.3,0,0.22,0.31,0.07,0.1,425.1070725,2.165852024,0.79802366
+0.16,0.22,0.29,0.16,0.17,0,438.8011137,3.858477387,1.01660159
+0.27,0,0.34,0.31,0.08,0,448.5728946,2.422814244,0.802378002
+0.17,0.21,0.21,0.25,0.16,0,436.961298,3.561703019,1.08299139
+0.24,0.12,0.26,0.32,0.06,0,451.0601178,2.041616095,1.018091383
+0.3,0,0.35,0.18,0.17,0,436.177002,4.006750528,0.662947313
+0,0.36,0.29,0.27,0.08,0,443.3685259,2.412987278,1.346
+0.12,0.21,0.31,0.33,0,0.03,422.0898274,0.920354014,1.170302789
+0,0.31,0.22,0.3,0,0.17,313.7802098,0.786912343,1.296
+0.1,0.23,0.35,0,0.15,0.17,327.803983,3.109105815,0.868286261
+0.14,0.23,0.32,0.2,0.11,0,447.3939985,2.836332285,1.079054559
+0,0.3,0.25,0.19,0.07,0.19,326.1397988,1.846426954,1.17
+0.31,0,0.32,0.17,0.2,0,391.9512959,4.549275105,0.669505348
+0.22,0.15,0.27,0.25,0.11,0,436.967869,2.899532144,0.996218701
+0.25,0.13,0.28,0.29,0,0.05,426.4898204,0.84847177,0.996880489
+0.23,0,0.32,0.26,0,0.19,401.519544,0.768496233,0.766645949
+0.19,0.15,0.29,0.32,0,0.05,404.2457425,0.987952392,1.056297893
+0.27,0,0.27,0.26,0.2,0,483.2164347,4.544728757,0.767616964
+0,0.3,0.24,0.28,0.18,0,492.0994642,4.242687882,1.26
+0.35,0,0.19,0.17,0.16,0.13,406.134562,3.219227307,0.684934249
+0.17,0.21,0.3,0.25,0.07,0,406.67483,2.135263501,1.084170089
+0.11,0.17,0.33,0.22,0,0.17,363.4549065,0.894561644,0.996720585
+0.38,0,0.28,0.17,0.17,0,387.5909487,3.756212357,0.677754259
+0,0.32,0.22,0.28,0.18,0,440.1281742,4.117540665,1.292
+0,0.39,0.16,0.29,0.16,0,436.9216758,3.564469337,1.414
+0.37,0,0.23,0.25,0,0.15,384.2922094,0.631977062,0.768333831
+0,0.32,0.31,0.31,0.06,0,405.7843704,2.111478941,1.322
+0.23,0.15,0.31,0.17,0,0.14,364.8985875,0.64387597,0.9106849
+0,0.4,0.2,0.19,0.11,0.1,358.526137,2.333813938,1.33
+0.21,0.12,0.19,0.25,0.11,0.12,408.5721848,2.489870735,0.943911986
+0,0.38,0.27,0.17,0,0.18,336.2541717,0.524039398,1.278
+0,0.3,0.27,0.23,0.2,0,400.7478785,4.538768738,1.21
